Move args parsing to new file
