+++ src/lib.rs
@@ -5,6 +5,7 @@ mod action;
 mod basic_types;
 mod expression;
 mod item;
+mod parse_args;
 mod pattern;
 mod value;
